test: test_gitlab is now sync agnostic
